JWT authentaction complete

+++ backend/controllers/goalControler.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalsModel')
+const User = require('../models/userModel')
 
 
 
@@ -12,7 +13,7 @@ const Goal = require('../models/goalsModel')
 // @access   Private
 const getGoals = asyncHandler(async (req, res) => {
 
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id })
   
   res.status(200).json(goals);
 });
@@ -27,7 +28,8 @@ const setGoal = asyncHandler(async(req, res) => {
     }
 
     const goal = await Goal.create({
-      text: req.body.text
+      text: req.body.text,
+      user: req.user.id
     })
 
   res.status(200).json(goal);
@@ -44,6 +46,20 @@ const updateGoal = asyncHandler(async(req, res) => {
     throw new Error('Goal not found')
   }
 
+  const user = await User.findById(req.user.id)
+
+  //check for user
+  if(!user){
+    res.status(401)
+    throw new Error ('User not Found')
+  }
+
+  //logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,{
     new:true,
   })
@@ -52,7 +68,7 @@ const updateGoal = asyncHandler(async(req, res) => {
 });
 
 // @desc     DeleteGoals
-// @route    elete / api / goals /:id
+// @route    delete / api / goals /:id
 // @access   Private
 const deleteGoal = asyncHandler(async(req, res) => {
 
@@ -61,6 +77,20 @@ const deleteGoal = asyncHandler(async(req, res) => {
   if(!goal){
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  //check for user
+  if(!user){
+    res.status(401)
+    throw new Error ('User not Found')
+  }
+
+  //logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
   
   await goal.deleteOne({id:req.params.id});
